fixed file naming to separate old + new versions of transcript counting in quantify
